// src/styles/types/common/BaseWidget.types.ts
export type BaseWidgetStyleType = {
  containerBorder: string;
  borderRadius: string;
  axisLabelToggleDimension: number;
  titleLabel: number;
  titleTagFontSize: number;
  iconGap: number;
  headerTop: number;
  headerBottom: number;
  headerLeft: number;
  headerRight: number;
  widgetTooltip: {
    backgroundColor: string;
    fontSize: number;
    color: string;
    padding: {
      top: number;
      right: number;
      bottom: number;
      left: number;
    };
  };
};